Add files via upload
This repository consists of the basic beginner level gcn code for a reddit user dataset. Considerin g the computational complexity, I have taken onlly the subset of the entire dataset.
I have uploaded the notebooks for:
   - Graph part
   - Machine learning Models 
   - Centrality measures
NOTE: The data dosent consists of the entire names of the edges and the associated weights. So this is only a beginner level project without the labels of the edges and associated weights.